Move TxnDetails into KitProvider dir
